Isolating the llg routines and containers.
